Extract file for RecognitionManager
